实现了浮点数的支持及其运算，主要通过扩增pl0虚拟机的代码实现。主要方式为首先将栈改为浮点数类型，然后为整数和浮点数分别实现不同的输入、输出pl0代码。

+++ mypl0.h
@@ -139,4 +139,4 @@ int vardeclaration(int* ptx, int lev, int* pdx);
 int constdeclaration(int* ptx, int lev, int* pdx);
 int position(char* idt, int tx);
 void enter(enum object k, int* ptx, int lev, int* pdx);
-int base(int l, int* s, int b);
+int base(int l, float* s, int b);
